Refactor: Simplify xp clear command to handle total deletion
